Upratovanie @ UI pre stahovanie zipka
